invoice-parser-step009/starting-code/backend/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
greenlet==3.0.1
email-validator==2.1.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.16.5
pydantic-settings==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
aiofiles==24.1.0
google-generativeai==0.8.3
langchain==0.3.15
langchain-google-genai==2.0.7
langchain-core==0.3.28
pillow==11.0.0
pypdf==5.1.0
pdf2image==1.17.0
tenacity==9.0.0